% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AfterBeau.R
\name{beautify_cate}
\alias{beautify_cate}
\title{Beautifying results (OR)}
\usage{
beautify_cate(data)
}
\arguments{
\item{data}{: results of logistic regression (in the form of OR and CI)}
}
\value{
Beautified results
}
\description{
Beautifying results (OR)
}
