% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarWrang.R
\name{cbind_dummy}
\alias{cbind_dummy}
\title{To be used with create_dummy(); creating dummies for multiple variables}
\usage{
cbind_dummy(masterdata, ind)
}
\arguments{
\item{masterdata}{: masterdata}

\item{ind}{: index of columns in masterdata that we want to create dummy variables}
}
\value{
Matrix of a series of combination of dummy variables
}
\description{
To be used with create_dummy(); creating dummies for multiple variables
}
