% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AfterBeau.R
\name{cont_beau}
\alias{cont_beau}
\title{Beautifying results of logistic regression (with normal regression output)}
\usage{
cont_beau(cont)
}
\arguments{
\item{cont}{: results of logistic regression with normal regression output}
}
\value{
beautified results
}
\description{
Beautifying results of logistic regression (with normal regression output)
}
