% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarWrang.R
\name{create_dummy}
\alias{create_dummy}
\title{A function to create dummy variables}
\usage{
create_dummy(column)
}
\arguments{
\item{column}{: a column of vectors, e.g. with values 1:6}
}
\value{
a matrix with number of columns equals number of vectors
}
\description{
A function to create dummy variables
}
