% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{dependent_var}
\alias{dependent_var}
\title{dependent_var}
\format{
A dataframe with 229 subjects (rows) and 18 variables (columns)
}
\usage{
dependent_var
}
\description{
A dataset containing all the dependent variables in a survey with regard to premature ejaculation
}
\keyword{datasets}
