% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugRelated.R
\name{drug2mult}
\alias{drug2mult}
\title{Function to calculate the multiplier for equivalent dosage (ref = fluoxetine)}
\usage{
drug2mult(x)
}
\arguments{
\item{x}{: indicating the type of antidepressants; 1: fluoxetine; 2: Sertraline; 3: Paroxetine; 4: Lexapro; 5: Cipram; 6: Fluvoxamine}
}
\value{
A dosage multiplier
}
\description{
Function to calculate the multiplier for equivalent dosage (ref = fluoxetine)
}
