% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugRelated.R
\name{drug_class}
\alias{drug_class}
\title{Classifying antidepressants into different drug classes}
\usage{
drug_class(x)
}
\arguments{
\item{x}{: drug name; case sensitive}
}
\value{
Drug class
}
\description{
Classifying antidepressants into different drug classes
}
