% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarWrang.R
\name{expand_var}
\alias{expand_var}
\title{Function to expand variables from a vector of characters}
\usage{
expand_var(masterdata, data, index, column)
}
\arguments{
\item{masterdata}{: masterdata; not much use, just to know the number of subjects}

\item{data}{: the dataframe containing the characters of variables}

\item{index}{: the index column in "data"}

\item{column}{: the column of characters in "data"}
}
\value{
A matrix with columns for different characters, and rows for different subjects, with values of 1 and 2
}
\description{
Function to expand variables from a vector of characters
}
