% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{foo}
\alias{foo}
\title{Frequencies of occurences}
\usage{
foo(data, col)
}
\arguments{
\item{data}{: the dataframe concerned}

\item{col}{: the column of the variable in the dataframe concerned}
}
\value{
a table presenting the frequencies of occurences
}
\description{
Frequencies of occurences
}
