% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reg.R
\name{lin_reg}
\alias{lin_reg}
\title{linear regression}
\usage{
lin_reg(datay, datax, j, i)
}
\arguments{
\item{datay}{: dataframe that contains the first variable, e.g. disease diagnosis status}

\item{datax}{: dataframe that contains the second variable}

\item{j}{: column number with respect to datay}

\item{i}{: column number with respect to datax}
}
\value{
Normal output of a regression
}
\description{
linear regression
}
