% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reg.R
\name{log_reg_spss_cont_adj0}
\alias{log_reg_spss_cont_adj0}
\title{Logistic regression with adjustment (continuous independent variable; including 0)}
\usage{
log_reg_spss_cont_adj0(datay, datax, j, i, adj)
}
\arguments{
\item{datay}{: dataframe that contains the first variable, e.g. disease diagnosis status}

\item{datax}{: dataframe that contains the second variable}

\item{j}{: column number with respect to datay}

\item{i}{: column number with respect to datax}

\item{adj}{: column number(s) in datax for adjustments}
}
\value{
Normal output of regression
}
\description{
Logistic regression with adjustment (continuous independent variable; including 0)
}
