% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reg.R
\name{log_reg_spss_kref}
\alias{log_reg_spss_kref}
\title{Logistic regression (categorical independent variable; specific referencing category)}
\usage{
log_reg_spss_kref(datay, datax, j, i, kref)
}
\arguments{
\item{datay}{: dataframe that contains the first variable, e.g. disease diagnosis status}

\item{datax}{: dataframe that contains the second variable}

\item{j}{: column number with respect to datay}

\item{i}{: column numbers with respect to datax}

\item{kref}{: specific referencing categories}
}
\value{
Odds ratio and CIs
}
\description{
Logistic regression (categorical independent variable; specific referencing category)
}
