% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reg.R
\name{log_reg_spss_kref_adj}
\alias{log_reg_spss_kref_adj}
\title{Logistic regression with adjustments (categorical independent variable; specific referencing category)}
\usage{
log_reg_spss_kref_adj(datay, datax, j, i, kref, adj)
}
\arguments{
\item{datay}{: dataframe that contains the first variable, e.g. disease diagnosis status}

\item{datax}{: dataframe that contains the second variable}

\item{j}{: column number with respect to datay}

\item{i}{: column number with respect to datax}

\item{kref}{: specific referencing categories}

\item{adj}{: column numbers for adjustments}
}
\value{
Odds ratio and CIs
}
\description{
Logistic regression with adjustments (categorical independent variable; specific referencing category)
}
