% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reg.R
\name{log_reg_spss_max}
\alias{log_reg_spss_max}
\title{Logistic regression (categorical independent variable; referenced to the category with most subjects)}
\usage{
log_reg_spss_max(datay, datax, j, i)
}
\arguments{
\item{datay}{: dataframe that contains the first variable, e.g. disease diagnosis status}

\item{datax}{: dataframe that contains the second variable}

\item{j}{: column number with respect to datay}

\item{i}{: column numbers with respect to datax}
}
\value{
Odds ratio and CIs
}
\description{
Logistic regression (categorical independent variable; referenced to the category with most subjects)
}
