% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{masterdata_A}
\alias{masterdata_A}
\title{masterdata_A}
\format{
A dataframe with 229 subjects (rows) and 103 variables (columns)
}
\usage{
masterdata_A
}
\description{
A dataset containing all the independent variables (non psychiatric related) in a survey with regard to premature ejaculation
}
\keyword{datasets}
