% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{masterdata_B}
\alias{masterdata_B}
\title{masterdata_B}
\format{
A dataframe with 229 subjects (rows) and 39 variables (columns)
}
\usage{
masterdata_B
}
\description{
A dataset containing all the independent variables (psychiatric related) in a survey with regard to premature ejaculation
}
\keyword{datasets}
