% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{medianiqr}
\alias{medianiqr}
\title{Median & IQR}
\usage{
medianiqr(data, col)
}
\arguments{
\item{data}{: the dataframe concerned}

\item{col}{: the column of the variable in the dataframe concerned}
}
\value{
Median and IQR of the concerned variable
}
\description{
Median & IQR
}
