% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AfterBeau.R
\name{merge_coef}
\alias{merge_coef}
\title{Merging the results (of logistic regression)}
\usage{
merge_coef(masterdata, path, number, key, ind)
}
\arguments{
\item{masterdata}{: masterdata}

\item{path}{: file path}

\item{number}{: file chapter number}

\item{key}{: identifier (e.g. ISSM / PEDT)}

\item{ind}{: desired variables for merging}
}
\value{
Merged results
}
\description{
Merging the results (of logistic regression)
}
