% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reg.R
\name{OR_CI}
\alias{OR_CI}
\title{Function calculating the confidence interval from the coefficient obtained by logistic regression}
\usage{
OR_CI(m)
}
\arguments{
\item{m}{: model calculated by glm()
return an array of odds ratio (LCI, OR, UCI)}
}
\description{
Function calculating the confidence interval from the coefficient obtained by logistic regression
}
