% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareDist.R
\name{serial_chisq_0}
\alias{serial_chisq_0}
\title{Serial chi-square tests (Including zeroes in the second variable)}
\usage{
serial_chisq_0(datay, datax, y, spec_x)
}
\arguments{
\item{datay}{: dataframe that contains the first variable, e.g. disease diagnosis status}

\item{datax}{: dataframe that contains the second to nth variables}

\item{y}{: column number with respect to datay}

\item{x}{: column numbers with respect to datax}
}
\value{
A matrix with all the resulted p-values
}
\description{
Serial chi-square tests (Including zeroes in the second variable)
}
