% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{serial_foo}
\alias{serial_foo}
\title{Frequencies of occurences for multiple variables}
\usage{
serial_foo(data, cols)
}
\arguments{
\item{data}{: the dataframe concerned}

\item{cols}{: the columns of the variable in the dataframe concerned}
}
\value{
a table presenting the frequencies of occurences for multiple variables
}
\description{
Frequencies of occurences for multiple variables
}
