% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reg.R
\name{serial_logregka}
\alias{serial_logregka}
\title{A Series of logistic regression with adjustments (categorical independent variable; specific referencing category)}
\usage{
serial_logregka(datay, datax, j, i, kref, adj)
}
\arguments{
\item{datay}{: dataframe that contains the first variable, e.g. disease diagnosis status}

\item{datax}{: dataframe that contains the second variable}

\item{j}{: column number with respect to datay}

\item{i}{: column numbers with respect to datax}

\item{kref}{: specific referencing categories}

\item{adj}{: column numbers for adjustments}
}
\value{
A beautified output for the series of regressions
}
\description{
A Series of logistic regression with adjustments (categorical independent variable; specific referencing category)
}
