% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reg.R
\name{serial_logregkacont}
\alias{serial_logregkacont}
\title{A series of logistic regression with adjustment (continuous independent variable)}
\usage{
serial_logregkacont(datay, datax, j, i, adj)
}
\arguments{
\item{datay}{: dataframe that contains the first variable, e.g. disease diagnosis status}

\item{datax}{: dataframe that contains the second variable}

\item{j}{: column number with respect to datay}

\item{i}{: column numbers with respect to datax}

\item{adj}{: column number(s) in datax for adjustments}
}
\value{
A beautified output for the series of regressions
}
\description{
A series of logistic regression with adjustment (continuous independent variable)
}
