% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{serial_miqr}
\alias{serial_miqr}
\title{A series of median & IQR}
\usage{
serial_miqr(data, cols)
}
\arguments{
\item{data}{: the dataframe concerned}

\item{cols}{: the columns of the variable in the dataframe concerned}
}
\value{
Median and IQR of the concerned variables
}
\description{
A series of median & IQR
}
