% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{serial_shapiro}
\alias{serial_shapiro}
\title{A series of shapiro test}
\usage{
serial_shapiro(data, cols)
}
\arguments{
\item{data}{: the dataframe concerned}

\item{cols}{: the columns of the variable in the dataframe concerned}
}
\value{
statistics and p-value of a series of shapiro tests
}
\description{
A series of shapiro test
}
