% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareDist.R
\name{serial_tab_0}
\alias{serial_tab_0}
\title{Creating a series of tables showing the distributions (Including zeroes in the second variable)}
\usage{
serial_tab_0(datay, datax, y, spec_x)
}
\arguments{
\item{datay}{: dataframe that contains the first variable, e.g. disease diagnosis status}

\item{datax}{: dataframe that contains the second to nth variable}

\item{y}{: column number with respect to datay}

\item{x}{: column numbers with respect to datax}
}
\value{
A series of tables with percentages
}
\description{
Creating a series of tables showing the distributions (Including zeroes in the second variable)
}
