% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIF.R
\name{serial_vif_cont_0}
\alias{serial_vif_cont_0}
\title{VIF for a series of logistic regression with adjustments (continuous var; including zero cont vars)}
\usage{
serial_vif_cont_0(datay, datax, j, i, adj)
}
\arguments{
\item{datay}{: dataframe that contains the first variable, e.g. disease diagnosis status}

\item{datax}{: dataframe that contains the second variable}

\item{j}{: column number with respect to datay}

\item{i}{: column numbers with respect to datax}

\item{adj}{: column numbers for adjustments}
}
\value{
VIF for variables in a series of logistic regression
}
\description{
VIF for a series of logistic regression with adjustments (continuous var; including zero cont vars)
}
