% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{serial_wilcox}
\alias{serial_wilcox}
\title{A series of wilcoxon test}
\usage{
serial_wilcox(datay, datax, j, is)
}
\arguments{
\item{datay}{: dataframe that contains the first variable, e.g. disease diagnosis status}

\item{datax}{: dataframe that contains the second to nth variables}

\item{j}{: column number with respect to datay}

\item{is}{: column numbers with respect to datax}
}
\value{
Statistics and p-values of a series of wilcoxon tests
}
\description{
A series of wilcoxon test
}
