% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{shapiro}
\alias{shapiro}
\title{Shapiro test}
\usage{
shapiro(data, col)
}
\arguments{
\item{data}{: the dataframe concerned}

\item{col}{: the column of the variable in the dataframe concerned}
}
\value{
statistics and p-value of shapiro test
}
\description{
Shapiro test
}
