% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareDist.R
\name{t.test02}
\alias{t.test02}
\title{Simple t test by statistics}
\usage{
\method{t}{test02}(n1, m1, sd1, n2, m2, sd2)
}
\arguments{
\item{n1}{: number of subjects in group 1}

\item{m1}{: mean of group 1}

\item{sd1}{: sd of group 1}

\item{n2}{: number of subjects in group 2}

\item{m2}{: mean of group 2}

\item{sd2}{: sd of group 2}
}
\value{
results of t test
}
\description{
Simple t test by statistics
}
