% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareDist.R
\name{t_test_ngo}
\alias{t_test_ngo}
\title{Simple t test}
\usage{
t_test_ngo(datay, datax, y, x)
}
\arguments{
\item{datay}{: dataframe that contains the first variable, e.g. disease diagnosis status}

\item{datax}{: dataframe that contains the second variable}

\item{y}{: column number with respect to datay}

\item{x}{: column numbers with respect to datax}
}
\value{
A matrix with the rounded t-test results
}
\description{
Simple t test
}
