% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareDist.R
\name{tab_ngo}
\alias{tab_ngo}
\title{Creating a table showing the distributions grouped by 2 variables}
\usage{
tab_ngo(datay, datax, y, x)
}
\arguments{
\item{datay}{: dataframe that contains the first variable, e.g. disease diagnosis status}

\item{datax}{: dataframe that contains the second variable, e.g. sex}

\item{y}{: column number with respect to datay}

\item{x}{: column number with respect to datax}
}
\value{
A table grouped by the 2 variables with percentages
}
\description{
Creating a table showing the distributions grouped by 2 variables
}
