% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareDist.R
\name{tab_norm}
\alias{tab_norm}
\title{Simple table without percentage}
\usage{
tab_norm(datay, datax, y, x)
}
\arguments{
\item{datay}{: dataframe that contains the first variable, e.g. disease diagnosis status}

\item{datax}{: dataframe that contains the second variable, e.g. sex}

\item{y}{: column number with respect to datay}

\item{x}{: column number with respect to datax}
}
\value{
A table grouped by the 2 variables
}
\description{
Simple table without percentage
}
