% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIF.R
\name{vif_adj}
\alias{vif_adj}
\title{VIF for logistic regression with adjustments (categorical independent variable; specific referencing category)}
\usage{
vif_adj(datay, datax, j, i, kref, adj)
}
\arguments{
\item{datay}{: dataframe that contains the first variable, e.g. disease diagnosis status}

\item{datax}{: dataframe that contains the second variable}

\item{j}{: column number with respect to datay}

\item{i}{: column number with respect to datax}

\item{kref}{: specific referencing categories}

\item{adj}{: column numbers for adjustments}
}
\value{
VIF for each variables
}
\description{
VIF for logistic regression with adjustments (categorical independent variable; specific referencing category)
}
