% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{wilcox}
\alias{wilcox}
\title{Wilcoxon test}
\usage{
wilcox(datay, datax, j, i)
}
\arguments{
\item{datay}{: dataframe that contains the first variable, e.g. disease diagnosis status}

\item{datax}{: dataframe that contains the second to nth variables}

\item{j}{: column number with respect to datay}

\item{i}{: column numbers with respect to datax}
}
\value{
Statistics and p-values of wilcoxon tests
}
\description{
Wilcoxon test
}
